From Alexandru Harsanyi                       Aug 3, 99 04:24:10 pm +0300
Hi,

I've made some enhancements to the dialog-add.c and dialog-edit.c files:

* my-xacc/src/gnome/dialog-add.c (gnc_ui_accWindow_tree_fill): the
account that is selected in the main window is selected in the
account creation dialog too.

* my-xacc/src/gnome/dialog-add.c
(gnc_ui_accWindow_create_callback): the account code is set for
the new account, if the user does not enter a string for the code,
xaccAccoutAutoCode is used. Thus each created accout will have a
code.

(gnc_ui_accWindow_create_callback): the function has been mostly
rewriten, it uses less variables, and the code for each account
component is in one place.

there's only one problem, though, xaccAccountBeginEdit is called,
but after that, the account editing might fail. There's no
xaccAccountRollbackEdit (yet?)

(make_valid_name): this function checks whether the name has some
non-whitespace characters in it. it trims the leading and trailing
whitespace and returns a new string. It is used as a check for
user entered data in gnc_ui_accWindow_create_callback, but it can
be used elsewhere too.

* my-xacc/src/gnome/dialog-edit.c and dialog-add.c: the  notes box has
got a scrollbar.

* my-xacc/src/gnome/dialog-edit.c: (editAccWindow) some accounts
do not have "securities", "descriptions" and "account codes"
set. In this case a check was added not to pass a NULL pointer to
gtk_entry_set_text. This removes annoying gtk warnings when gnucash
runs.

* the gnc_ui_accWindow_tree_expand function has been removed from
dialog-add.c since it is not used.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@1887 57a11ea4-9604-0410-9ed3-97b8803252fd