Install.sh: Add set_env_or_die, use it for $INTLTOOL_PERL.
This function is similar to add_to_env, but makes the script fail if it
would effectively overwrite a non-empty variable.  This is not the case
if the new value was already a part of the variables contents, no matter
whether from the beginning or after an add_to_env call.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@15585 57a11ea4-9604-0410-9ed3-97b8803252fd