Add support for (de)compressing in a thread instead of fork.
Use GThreading system when available, but depend on it on Windows.
Gzipped data files are still loaded by libxml2.  On write, again create
a pipe p and a thread t that gzwrites bytes read from p to a given file.
After the other thread has finished writing to p, it waits for t to
return to avoid loss of data.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@15748 57a11ea4-9604-0410-9ed3-97b8803252fd