Play nicer with recent releases of glib that have tightened what's considered a valid key in a key/value file. This change will cause gnucash to rewrite any opened state file into a new version where all key names fit the new glib requirements.
This code was originally committed to trunk as #15458 on 1/28/07, and
then backed out with #15461 on 1/29/07.  Backward compatibility
support for this code was committed to the 2.0 branch in #15460 and
was released as part of 2.0.5.  Now that 2.0.5 has now been out for
six weeks, this commit puts the original changes back into trunk.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@15787 57a11ea4-9604-0410-9ed3-97b8803252fd