Add optional two-column display of income statement and reversing the income/expense order.
This patch implements a two-column income statement. I have sent this to the
original person who requested it on IRC, but haven't heard back as to whether
it actually solves his problem. Additionally, the patch includes support for reversing the order of income/expense
tables in the report. It defaults to the standard income/expense, but can be
reversed to show expense/income (bad news first?) and is supported by either
one or two-column mode.

Patch by Andrew Sackville-West, bug#488004

Signed-off-by: Christian Stimming <stimming@tuhh.de>

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@16569 57a11ea4-9604-0410-9ed3-97b8803252fd