Implement a Report GUID.
This is an internal representation of the report. It frees up the various name
fields for translation/changes without concern about it breaking reports*. Each
report is now assigned a GUID that refers to that specific report. All
reference to the report is now done with this GUID.

Support is included for using existing open or saved reports of the old
reference-by-name type. The user is warned of the existence of saved reports
without a GUID. Support is also provided to allow reports saved or left open
using the new report-guid reference to be accessed in 2.2.3(?) versions in case
user downgrades.

IMPORTANT: All saved or open reports created using these changes *will* *cause*
*application* *crashes* if accessed in versions prior to r16804. Earlier
versions have neither the new functions nor the report-record fields
implemented here. It is a one-way trip from pre-r16804 to here. I hope that is
clear enough ;)

* going forward only. name changes will still break non-report-guid saved or
open reports.


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@16805 57a11ea4-9604-0410-9ed3-97b8803252fd