QIF Import: Fix support for multi-byte account separators.
In doing so, a number of reusable Scheme string manipulation procedures were written and placed in string.scm. These are now available to all Scheme code by automatic inclusion in main.scm.

The new Scheme procedures are:
gnc:string-rcontains    (a variation on string-contains)
gnc:substring-count     (a variation on string-count)
gnc:substring-split     (a variation on string-split)
gnc:substring-replace   (search/replace a substring)
gnc:string-replace-char (search/replace a character)
gnc:string-delete-chars (delete a variety of characters)

Finally, the custom version of string-split was removed because Guile 1.4 is no longer supported and later versions come with this procedure.
BP


git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17191 57a11ea4-9604-0410-9ed3-97b8803252fd