Bug #549595, #552306: Correct retval of gnc_int_link_or_make_backup(), add ENOSYS and ENOTSUP.
The return value of link(2) on files residing on sshfs will be ENOSYS, on
network hfsplus file systems on mac ENOTSUP, so add them to the list of those
errnos that trigger copy_file as fallback in gnc_int_link_or_make_backup() or
are allowed in gnc_file_be_get_file_lock() showing that hardlinks are not
supported.

To avoid silent data loss in the case of an unexpected errno in
gnc_int_link_or_make_backup(), correctly set the backend error.

Patches from Micha Lenk and Boris Zbarsky.
BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17524 57a11ea4-9604-0410-9ed3-97b8803252fd