Bug #551858: Add Job Report for the business module
The contributor writes:
I need a report which shows me all invoices of a job. This is similar to the
owner-report, e.g. Customer Report. I took the file of owner-report.scm,
changed it so it does the desired and saved it as job-report.scm. I also had to
patch business-core.i to export the right symbols and business-urls.c to have
access to a link to the Job in the header. A patch is attached and I would be
very pleased if this could make it into gnucash. It could be possible to unify
owner-report and job-report, but I didn't put too much effort in it.

Patch by Stefan Wolf.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17664 57a11ea4-9604-0410-9ed3-97b8803252fd