Multicurrency business features: Handle invoice/bill line items
This patch is the last in the multicurrency set : it deals with
conversions where individual invoice/bill line item accounts are
different from the "owner" currency (aka, the invoice/bill currency).

The original plan was to also have a checkbox to allow the user to use
an account that isn't usually in the owner currency, in the owner
currency, but that's more ornate and requires more time and code than I
have time for right now, so, it's a feature for future :)

I tested out "invoice currency matches the default books currency and
has some accounts that use other currencies" and "invoice currency does
not match default currency and has some stuff that does and some stuff
that doesn't".. it could probably stand to also be tested by some other
folks, but what I tried worked :)

Patch by Jamie Campbell.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17719 57a11ea4-9604-0410-9ed3-97b8803252fd