Fix Bug 559772 – SQL backend not non-ascii-safe
Sqlite3 uses utf8 encoding for all char fields, so it is non-ascii-safe.  
For postgresql, the default encoding can be set on a per-db basis.  
Since the database is not created by gnucash (the tables are, but not 
the database), it is for the user to set utf8 encoding when the database 
is created.  For mysql, a default encoding can be set on a per-db, 
per-table or per-field basis.  Since there are char fields which do not 
need to be utf8 (e.g. guids), encoding is set on a per-field basis.



git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17780 57a11ea4-9604-0410-9ed3-97b8803252fd