Bug #564450: HUF currency can be divided into 100 "fillers" even though there are no coins for it
In Hungary the smallest Forint (HUF) coin is 5 forints, but formally we had 1,2
forint coins, and dividing FILLER coins (like cents for US, EUR, ...). They are
simply not circulated in cash usage, but the banking system still using it.
Such as the smallest transferable amount is 0.01 FORINTs (1 FILLER), and with
business it is normal to use it (as result of foreign exchange, incoming
interest rate, or many other reason).

Patch by Frank H. Ellenberger.
BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17849 57a11ea4-9604-0410-9ed3-97b8803252fd