Bug #569734: Give the template root account a name
The template root account created as part of the scheduled transaction
infrastructure should have a name.  Various parts of the Gnucash engine get
upset if an account doesn't have a name.
Perhaps this name should be localized, but the name of the main root account is
not localized.

Patch by Mike Alexander.
BP

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@17867 57a11ea4-9604-0410-9ed3-97b8803252fd