I18n improvements: Don't mark single characters for translation.
Also, all of the split sentences should be refactored into printf-like
strings where the variable words must be inserted through %s etc, but I'm
too lazy to do this for now. Also, this file us/taxtxf.scm is used in the
U.S. *only*, so it is fine to not translate it anyway.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18049 57a11ea4-9604-0410-9ed3-97b8803252fd