Partly fix broken data file backward compatibility where SX recurrence would disappear.
In r17725 and r17751, I introduced the extra XML child element
<recurrence:weekend_adj>, but this means a gnucash-2.2.x cannot read
the SX recurrence of a >=2.3.x file anymore, which is bad. In order
to improve this broken backward compatibility for at least the common
case, we don't write out this XML element as long as it is only "none".

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18376 57a11ea4-9604-0410-9ed3-97b8803252fd