Bug #593070: Budget Report: Omit accounts with no budget and zero balance
Patch contributed by Luke Duncan:

I would like to have an option for the budget report that allows it to omit
accounts that have zero balance and no budget value specified over the entire
budget term. This avoids showing entire rows of blank cells without the user
having to manually select which accounts have a budget value or balance for the
term of the budget. The user can select all expense and income accounts and it
only shows the accounts that have actual (nonzero) data to display.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18394 57a11ea4-9604-0410-9ed3-97b8803252fd