Bug #540452: Remove inline from function declaration/definition because it contradicts the C99 standard.
Patch by Halton Huo: Remove inline from qofmath128.c and qofmath128.h.

Cstim notes: The way the code is written (.h/.c) we should not have
any "inline" there. If the functions are intended to be used inline
only (and they are, in gnc-numeric.c), the definitions should be moved
into gnc-numeric.c anyway and the declarations removed.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18416 57a11ea4-9604-0410-9ed3-97b8803252fd