Bug #601861: Fix mis-captioning in button labels
The keyboard shortcut hints in some dialogs are miscaptioned. This means that
in a save request dialog, for example, the button caption "Save Transaction"
spuriously reads "_Save Transaction" instead of "Save Transaction" with the "S" of "Save"
being underlined to mark the proper key on the keyboard to press in order to invoke
saving of the transaction.

Seems like new gtk-version is stricter and adds mnemonics only if told so.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18419 57a11ea4-9604-0410-9ed3-97b8803252fd