Bug #605415: Relax limit on scheduled days in the future
Patch by James Raehl: Expand scheduled transaction date limit by 366 days

Usually, scheduled transactions are limited to 365 days in the future.  This
patch doubles that limit plus adds a day for leap years.  The patch supports a
method for separating transactions that are not yet posted at the bank, by
assigning an initial date one year in the future.  Under this method, scheduled
transactions and their reminders are one year in the future, since they have
not happened yet.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18537 57a11ea4-9604-0410-9ed3-97b8803252fd