Bug #606077: Fix uncommon configure option --program-suffix.
Patch by Peter Selinger:

OK, the attached patch fixes part of this problem. Specifically, it applies the
specified transform (from ./configure option --program-suffix,
--program-prefix, or --program-transform-name) to the files

src/bin/gnucash-valgrind
src/bin/gnucash-gdb
src/bin/gnucash

This is done in the Makefile at compile time.

In the end, I am not sure why the GnuCash configure.in provides the
--program-suffix mechanism at all, since apparently it hasn't been used in 10
years. According to info autoconf, these are supposed to be enabled by the
macro AC_ARG_PROGRAM, but I don't see this in configure.in. So perhaps it has
been enabled by accident and shouldn't really be there.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18572 57a11ea4-9604-0410-9ed3-97b8803252fd