MSVC compatiblity: Use a char* pointer for the memcpy() input argument.
This is required by MSVC because we do some pointer arithmetic
in the memcpy() argument, but in order to do this, MSVC wants to
know the pointed-to type of the pointer because pointer arithmetic
increases the pointer not by a number a bytes but a number of
sizeof(type). MSVC thinks for void* it doesn't count bytes. We
achieve the desired effect by using a char* pointer so that bytes
are counted.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18752 57a11ea4-9604-0410-9ed3-97b8803252fd