Fixup binreloc, which wasn't working correctly on OSX
Most changes are to rename the functions named gbr_* to gnc_gbr_*, so
that our modified binreloc (to provide mac bundle support) doesn't
collide with binreloc implementations in other libraries. (The instant
case was in libenchant.)

Also moved the IGE_MAC_CFLAGS and IGE_MAC_LIBS lines from
engine/Makefile.am to core-utils/Makefile.am. This was missed when
binreloc was moved.

And last, there was a minor error in the original OSX binreloc patch:
At line 68, #elif MAC_INTEGRATION should be (and now is) #elif defined
MAC_INTEGRATION.



git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@18940 57a11ea4-9604-0410-9ed3-97b8803252fd