Bug #615168: Make sure that printf is available whereever it is used
So, the right thing is for each file to be in its own module. Barring
that it's tough to see what the environment is in a given module. So to
be safe, everywhere there's a printf, make sure the (gnucash printf)
module is duly included.

Patch by Andy Wingo.

Conflicts:

	src/business/business-reports/easy-invoice.scm
	src/business/business-reports/fancy-invoice.scm
	src/business/business-reports/invoice.scm
	src/business/business-reports/job-report.scm
	src/report/standard-reports/budget-income-statement.scm
	src/report/standard-reports/income-statement.scm
	src/report/standard-reports/trial-balance.scm

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@19067 57a11ea4-9604-0410-9ed3-97b8803252fd