Bug 612337: Add file save to hierarchy druid when the new file druid completes.
Patch by "Bob": I thought I would give this a try and come up with a patch to use the file save
as function after the after pressing the apply button. Not sure if it is the
best way but seems to work on my linux box and on windows without any errors.

I have changed the text in the account.glade file to reflect the firing of the
save as and also increased the size of the window to give more room for the
'Category Description' to stop scroll bars.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@19377 57a11ea4-9604-0410-9ed3-97b8803252fd