Bug #589954: No blank split after creating a new account "on the fly" while editing a transaction.
Cause was that gnc_split_register_get_account_by_name() turned of "full_refresh" if the account didn't already exist and turned it back on only if the name of the account actually created in the new account dialog box differed from the one originally entered in the transaction.

Since "full_refresh" isn't manipulated much, that effectively shut down the ledger for further editing. Probably not what the original author intended!



git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@19847 57a11ea4-9604-0410-9ed3-97b8803252fd