Bug #636402: Add test case for DBI error: 1: table sqlite_sequence may not be dropped
Patch by Mark Jenkins:

The attached patch adds a test program demonstrating the problem in C, built in
src/optional/python-bindings
It pretty much shows the same calls that are being made to the C by the python
bindings (including initialization) and the example sqlite3 python script.
Running this small C program produces the same result.

(The program is built in check_PROGRAMS, not bin_PROGRAMS, to avoid having it installed.)

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@20023 57a11ea4-9604-0410-9ed3-97b8803252fd