Bug #638543: Validate counter format strings before using them.
Patch by Matthijs Kooijman:

The validation function is a very simple "parser" that simply checks for
a single gint64 conversion specification, allowing all modifiers and
flags that printf(3) specifies (except for the * width and precision,
which need an extra argument).

The validation function returns an error message that is used to log a
warning and can be used by the GUI later on.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@20056 57a11ea4-9604-0410-9ed3-97b8803252fd