Bug #638225: Sort when saving as XML
Patch by Jim Radford (with code beautified and re-indented by myself):

The attached patches sort the slots, lots, book accounts, bill terms,
customers, employees, entries, invoices, jobs, orders, tax tables and
vendors before saving them to the GnuCash XML file.

This is an attempt to make saves more idempotent thereby facilitating
the use of a revision control system on the GnuCash XML files.

With these patches most of the needless and seemingly random churn is
gone and I can add or remove a transaction and expect
there to be no unrelated changes to the GnuCash file.

I've been using and refining this patches for the last few years, so it has
received quite a bit of testing.

David Fraser adds: Without specific testing, I'm using this on an average-sized gnucash file
(5.7MB) without noticing any particular slowdown in saving, but a wonderful
reduction in diffs when comparing changes.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@20067 57a11ea4-9604-0410-9ed3-97b8803252fd