Add a read-only attribute to QofBook and check of same to qof_instance_begin_edit. Add qof_session_safe_save. Add unit tests for these changes.
QofBook read-only attribute is to prevent starting an edit when the 
version of Gnucash which created a dataset is newer than the one 
currently running. This is to prevent writing records of an old format 
into an existing data set. The user can use File>Save As to create a new 
QofSession which will write a completely new dataset in the current 
format. This is an important safety feature for future versions of 
Gnucash which will not bring all of a dataset into memory, instead 
loading only what is immediately needed.

Safe-save is the Qof access to allow writing out a dataset to an 
existing server database (i.e., postgresql or mysql) while maintaining 
the ability to both transaction-protect saving individual records while 
maintaining the ability to rollback the entire save if something goes 
wrong.

The unit test framework is added to enable testing these components. 
More tests will be added in a later commit.



git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/gnucash/trunk@20103 57a11ea4-9604-0410-9ed3-97b8803252fd